HDPI-1367: code clean up
