+++ build.gradle
@@ -432,7 +432,7 @@ tasks.named('test') {
 
 tasks.withType(CftlibExec).configureEach {
   group = 'ccd tasks'
-  authMode = AuthMode.AAT
+  authMode = AuthMode.Local
 
   //Attach mockito agent for inline mocking to remove warning in logs
   classpath += configurations.mockitoAgentConfig
